refactor: ui 컴포넌트 패키지 변경
- `payments` 패키지에서 `component` 패키지로 이동
