+++ app/src/androidTest/java/woowacourse/payments/ui/CardNumberTextFieldTest.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.payments.CardNumberTextField
+import woowacourse.payments.ui.component.CardNumberTextField
 
 class CardNumberTextFieldTest {
     @get:Rule
