+++ app/src/androidTest/java/woowacourse/payments/ui/CardPasswordTextFieldTest.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.payments.CardPasswordTextField
+import woowacourse.payments.ui.component.CardPasswordTextField
 
 class CardPasswordTextFieldTest {
     @get:Rule
